move business logic to the service layer
