+++ src/constants/company.ts
@@ -1,0 +1,11 @@
+// Info: (20260701 - Julian) 企業別名映射表
+export const COMPANY_ALIASES: Record<string, string[]> = {
+  台灣水泥股份有限公司: ["台泥", "台灣水泥"],
+  亞洲水泥股份有限公司: ["亞泥", "亞洲水泥"],
+  台灣積體電路製造股份有限公司: ["台積電", "台積", "tsmc"],
+  王道商業銀行股份有限公司: ["王道", "王道銀行"],
+  精誠資訊股份有限公司: ["精誠", "精誠資訊"],
+  臺灣企銀: ["台企銀", "臺企銀"],
+  遠傳電信股份有限公司: ["遠傳", "遠傳電信"],
+  長榮海運股份有限公司: ["長榮", "長榮海運"],
+};
